+++ projects/ngx-uploader-directive/CHANGELOG.md
@@ -4,14 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.4] - 2020-06-15
 
 ### Fixed
 - Fixed error event emittion.
 - Fixed events emittion on invalid request id.
 
 ### Changed
-- Readme.
+- Readme
 - Changed error response to HttpErrorResponse
 - On rejection Http Error response will be returned
 
@@ -44,7 +44,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added File select directive
 - Uploading files in a single request
 
-[Unreleased]: https://github.com/jayprajapati857/ngx-uploader-directive/compare/1.1.3...HEAD
+[1.1.4]: https://github.com/jayprajapati857/ngx-uploader-directive/compare/1.1.4...1.1.3
 [1.1.3]: https://github.com/jayprajapati857/ngx-uploader-directive/compare/1.1.2...1.1.3
 [1.1.2]: https://github.com/jayprajapati857/ngx-uploader-directive/compare/1.1.0...1.1.2
 [1.1.0]: https://github.com/jayprajapati857/ngx-uploader-directive/compare/1.0.0...1.1.0